function rename

+++ test/simple.c
@@ -12,7 +12,7 @@ TEST_FIXTURE_TEAREDOWN(simple)
 }
 TEST_PARAMETERIZED_DEFINE(simple, size_t, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
 
-static int s_sum;
+static size_t s_sum;
 
 TEST_F(simple, fixture)
 {
